Moved SuddenPlusCalculator to helpers package
