+++ app/src/main/java/com/duraz0rz/suddenpluscalculator/activities/SuddenPlusTableActivity.kt
@@ -7,7 +7,7 @@ import android.widget.TableLayout
 import android.widget.TableRow
 import android.widget.TextView
 import com.duraz0rz.suddenpluscalculator.R
-import com.duraz0rz.suddenpluscalculator.SuddenPlusCalculator
+import com.duraz0rz.suddenpluscalculator.helpers.SuddenPlusCalculator
 import com.duraz0rz.suddenpluscalculator.dataClasses.SuddenPlusValue
 
 class SuddenPlusTableActivity : AppCompatActivity() {
