+++ app/src/test/java/com/duraz0rz/suddenpluscalculator/activities/SuddenPlusTableActivityTest.kt
@@ -5,7 +5,7 @@ import android.widget.TableLayout
 import android.widget.TableRow
 import android.widget.TextView
 import com.duraz0rz.suddenpluscalculator.R
-import com.duraz0rz.suddenpluscalculator.SuddenPlusCalculator
+import com.duraz0rz.suddenpluscalculator.helpers.SuddenPlusCalculator
 import com.duraz0rz.suddenpluscalculator.dataClasses.SuddenPlusValue
 import com.natpryce.hamkrest.assertion.assertThat
 import com.natpryce.hamkrest.equalTo
