+++ app/src/test/java/com/duraz0rz/suddenpluscalculator/helpers/SuddenPlusCalculatorTest.kt
@@ -1,6 +1,7 @@
-package com.duraz0rz.suddenpluscalculator
+package com.duraz0rz.suddenpluscalculator.helpers
 
 import com.duraz0rz.suddenpluscalculator.dataClasses.SuddenPlusValue
+import com.duraz0rz.suddenpluscalculator.helpers.SuddenPlusCalculator
 import com.natpryce.hamkrest.assertion.assertThat
 import com.natpryce.hamkrest.closeTo
 import com.natpryce.hamkrest.equalTo
